Created new file for socket events
